url파일 생성
- ignore시키기 위하여 url파일 분리
