+++ dormitoryFamiles/.gitignore
@@ -115,6 +115,7 @@ iOSInjectionProject/
 
 ### private ###
 *Token.swift
+*Url.swift
 
 
 # End of https://www.gitignore.io/api/xcode,swift,cocoapods
